Change "debug" config dynamically
