+++ dredis/config.py
@@ -1,6 +1,8 @@
 import fnmatch
+import logging
 
 from dredis.exceptions import DredisError
+from dredis.utils import setup_logging
 
 TRUE = 'true'
 FALSE = 'false'
@@ -25,6 +27,14 @@ def get_all(pattern):
 
 def set(option, value):
     if option in _SERVER_CONFIG:
+        if option == 'debug':
+            value = _validate_bool(option, value)
+            if value == TRUE:
+                setup_logging(logging.DEBUG)
+            else:
+                setup_logging(logging.INFO)
+        elif option == 'readonly':
+            value = _validate_bool(option, value)
         _SERVER_CONFIG[option] = value
     else:
         raise DredisError('Unsupported CONFIG parameter: {}'.format(option))
@@ -32,3 +42,9 @@ def set(option, value):
 
 def get(option):
     return _SERVER_CONFIG[option]
+
+
+def _validate_bool(option, value):
+    if value.lower() not in (TRUE, FALSE):
+        raise DredisError("Invalid argument '{}' for CONFIG SET '{}'".format(value, option))
+    return value.lower()
